Backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg
psycopg2-binary
python-dotenv
drf-spectacular
django-cors-headers